Make wsprsim.c accept non-integer SNR
git-svn-id: svn+ssh://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@8392 ab8295b8-cf94-4d9e-aec4-7959e3be5d79

+++ lib/wsprd/wsprsim.c
@@ -151,7 +151,8 @@ int main(int argc, char *argv[])
                 writec2=1;
                 break;
             case 's':
-                snr = (float)atoi(optarg);
+//                snr = (float)atoi(optarg);
+                snr = atof(optarg);
                 break;
         }
     }
